Implement faster DecodeURI/DecodeURIComponent
This offers two improvements over the previous decoding algorithm:

1. It uses a reworked version of Bjoern Hoehrmann's DFA Unicode Decoder,
which handles all edge cases (overlong, invalid continuation, etc) by
design.
2. It preallocates the decoded string to the encoded string. This allows
us to drop down into `std::find` and `std::move`, which is vastly faster
than individual `::push_back`s.

- - -

The DFA itself is unnecessary for the Speed benefits, but it just makes
decoding so much nicer. Essentially, reworks a standard FSM (imagine an
array of structs) and flattens it out into a single dimension array.
Using Table 3-7 of the Unicode 10.0.0 standard (page 126 of
http://www.unicode.org/versions/Unicode10.0.0/ch03.pdf), we can nicely
map all bytes into one of 12 character classes:

0. 0x00-0x7F
1. 0x80-0x8F (split from general continuation because this range is not
  valid after a 0xF0 leading byte)
2. 0x90-0x9F (split from general continuation because this range is not
  valid after a 0xE0 nor a 0xF4 leading byte)
3. 0xA0-0xBF (the rest of the continuation range)
4. 0xC0-0xC1, 0xF5-0xFF (the joined range of invalid bytes, notice this
  includes 255 which we use as a known bad byte during hex-to-int decoding)
5. 0xC2-0xDF (leading bytes which require any continuation byte
  afterwards)
6. 0xE0 (leading byte which requires a 0xA0-0xBF afterwards then any
  continuation byte after that)
7. 0xE1-0xEC, 0xEE-0xEF (leading bytes which requires any continuation
  afterwards then any continuation byte after that)
8. 0xED (leading byte which requires a 0x80-0x9F afterwards then any
  continuation byte after that)
9. 0xF1-F3 (leading bytes which requires any continuation byte
  afterwards then any continuation byte then any continuation byte)
10. 0xF0 (leading bytes which requires a 0x90-0xBF afterwards then any
  continuation byte then any continuation byte)
11. 0xF4 (leading bytes which requires a 0x80-0x8F afterwards then any
  continuation byte then any continuation byte)

Note that 0xF0 and 0xF1-0xF3 were swapped so that less bytes were needed
to represent the transition state ("9, 10, 10, 10" vs "10, 9, 9, 9").

Using these 12 classes as "transitions", we can map from one state to
the next. Each state is defined as some multiple of 12, so that we're
always start at the 0th column of each row of the FSM. From each state,
we add the transition and get a index of the new row the FSM is
entering.

If at any point we encounter a bad byte, the state + bad-byte-transition
is guaranteed to map us into the first row of the FSM (which contains no
valid exiting transitions).

The key differences from Bjoern's original (or his self-modified) DFA is
the "bad" state is now mapped to 0 (or the first row of the FSM) instead
of 12 (the second row). This saves ~50 bytes when gzipping, and also
speeds up determining if a string is properly encoded (see his sample
code at http://bjoern.hoehrmann.de/utf-8/decoder/dfa/#performance).

Finally, I've replace his ternary check with an array access, to make
the algorithm branchless. This places a requirement on the caller to 0
out the codepoint between successful decodings, which it could always
have done because it's already branching.